a lot of changes. add google maps, change the design a bit. add faceb…

+++ app/src/main/java/lk/damithab/curenex/fragment/AddressFragment.java
@@ -1,5 +1,9 @@
 package lk.damithab.curenex.fragment;
 
+import android.app.Activity;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
+import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.activity.OnBackPressedCallback;
@@ -11,6 +15,7 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.auth.FirebaseAuth;
@@ -23,6 +28,7 @@ import java.util.List;
 import lk.damithab.curenex.R;
 import lk.damithab.curenex.adapter.AddressAdapter;
 import lk.damithab.curenex.databinding.FragmentAddressBinding;
+import lk.damithab.curenex.dialog.CustomAlertDialog;
 import lk.damithab.curenex.dialog.ToastDialog;
 import lk.damithab.curenex.model.Address;
 
@@ -56,16 +62,16 @@ public class AddressFragment extends Fragment {
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        getActivity().getOnBackPressedDispatcher().addCallback(getViewLifecycleOwner(), new OnBackPressedCallback(true) {
-            @Override
-            public void handleOnBackPressed() {
-                requireActivity().getSupportFragmentManager().popBackStack();
-            }
-        });
+//        getActivity().getOnBackPressedDispatcher().addCallback(getViewLifecycleOwner(), new OnBackPressedCallback(true) {
+//            @Override
+//            public void handleOnBackPressed() {
+//                requireActivity().getSupportFragmentManager().popBackStack();
+//            }
+//        });
 
         binding.addNewAddressBtn.setOnClickListener(v->{
             AddAddressFragment addAddressFragment = new AddAddressFragment();
-            getParentFragmentManager().beginTransaction().replace(R.id.navContainerView, addAddressFragment)
+            getParentFragmentManager().beginTransaction().replace(R.id.address_fragment_container, addAddressFragment)
                     .addToBackStack(null)
                     .commit();
         });
@@ -82,26 +88,48 @@ public class AddressFragment extends Fragment {
 
                             binding.userAddressRecycler.setLayoutManager(new LinearLayoutManager(getContext(), LinearLayoutManager.VERTICAL, false));
                             AddressAdapter adapter = new AddressAdapter(addressList, address->{
-                                Bundle result = new Bundle();
-                                result.putSerializable("selectedAddress", address); // Ensure Address class implements Serializable
-
-                                // Set the result with a unique key
-                                getParentFragmentManager().setFragmentResult("addressRequest", result);
-                                getParentFragmentManager().popBackStack();
+//                                Bundle result = new Bundle();
+//                                result.putSerializable("selectedAddress", address); // Ensure Address class implements Serializable
+//
+//                                // Set the result with a unique key
+//                                getParentFragmentManager().setFragmentResult("addressRequest", result);
+//                                getParentFragmentManager().popBackStack();
+                                Intent resultIntent = new Intent();
+                                resultIntent.putExtra("selectedAddress", address);
+                                requireActivity().setResult(Activity.RESULT_OK, resultIntent);
+                                requireActivity().finish();
                             });
 
                             adapter.setOnRemoveListener(position -> {
                                 String docId = addressList.get(position).getAddressId();
-                                db.collection("address").document(docId).delete()
-                                        .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                            @Override
-                                            public void onSuccess(Void unused) {
-                                                new ToastDialog(getActivity().getSupportFragmentManager(), "Address removed successfully!");
-                                                addressList.remove(position);
-                                                adapter.notifyItemRemoved(position);
-                                                adapter.notifyItemRangeChanged(position, addressList.size());
-                                            }
-                                        });
+                                new CustomAlertDialog(getContext())
+                                        .setTitle("Confirmation Message")
+                                        .setMessage("Are you sure you want to delete this address?")
+                                        .setPositiveButton("Remove", v -> {
+                                            db.collection("address").document(docId).delete()
+                                                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                                        @Override
+                                                        public void onSuccess(Void unused) {
+                                                            new ToastDialog(getActivity().getSupportFragmentManager(), "Address removed successfully!");
+                                                            addressList.remove(position);
+                                                            adapter.notifyItemRemoved(position);
+                                                            adapter.notifyItemRangeChanged(position, addressList.size());
+                                                        }
+                                                    });
+                                        })
+                                        .setNegativeButton()
+                                        .show();
+//                                new AlertDialog.Builder(getActivity())
+//                                        .setTitle("")
+//                                        .setMessage("")
+//                                        .setPositiveButton("OK", new DialogInterface.OnClickListener() {
+//                                            @Override
+//                                            public void onClick(DialogInterface dialogInterface, int i) {
+//
+//                                            }
+//                                        })
+//                                        .setNegativeButton("No", null)
+//                                        .show();
                             });
 
                             adapter.setOnEditListener(address -> {
@@ -111,7 +139,7 @@ public class AddressFragment extends Fragment {
                                 addAddressFragment.setArguments(bundle);
 
                                 getParentFragmentManager().beginTransaction()
-                                        .replace(R.id.navContainerView,addAddressFragment)
+                                        .replace(R.id.address_fragment_container,addAddressFragment)
                                         .addToBackStack(null)
                                         .commit();
                             });
@@ -127,14 +155,14 @@ public class AddressFragment extends Fragment {
     @Override
     public void onResume() {
         super.onResume();
-        getActivity().findViewById(R.id.bottomNavigationView).setVisibility(View.GONE);
-        getActivity().findViewById(R.id.main_toolbar).setVisibility(View.GONE);
+//        getActivity().findViewById(R.id.bottomNavigationView).setVisibility(View.GONE);
+//        getActivity().findViewById(R.id.main_toolbar).setVisibility(View.GONE);
     }
 
     @Override
     public void onStop() {
         super.onStop();
-        getActivity().findViewById(R.id.bottomNavigationView).setVisibility(View.VISIBLE);
-        getActivity().findViewById(R.id.main_toolbar).setVisibility(View.VISIBLE);
+//        getActivity().findViewById(R.id.bottomNavigationView).setVisibility(View.VISIBLE);
+//        getActivity().findViewById(R.id.main_toolbar).setVisibility(View.VISIBLE);
     }
 }
